some refactoring, renaming
